var types, build errors fix.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@5792 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_ivr_async.c
@@ -303,7 +303,7 @@ SWITCH_DECLARE(switch_status_t) switch_ivr_record_session(switch_core_session_t 
 	switch_status_t status;
 	time_t to = 0;
 	switch_media_bug_flag_t flags = SMBF_READ_STREAM | SMBF_WRITE_STREAM;
-	int channels;
+	uint8_t channels;
 	channel = switch_core_session_get_channel(session);
 	assert(channel != NULL);
